Move deployment wizard extension points into dedicated file
