+++ packages/model-serving/src/components/deploymentWizard/deploying/useDeployMethod.ts
@@ -2,12 +2,12 @@ import React from 'react';
 import type { ResolvedExtension } from '@openshift/dynamic-plugin-sdk';
 import { useResolvedExtensions } from '@odh-dashboard/plugin-core';
 import type { WizardFormData } from '../types';
+import type { Deployment } from '../../../../extension-points';
 import {
   DeploymentAssemblyResources,
   isModelServingDeploy,
-  type Deployment,
   type ModelServingDeploy,
-} from '../../../../extension-points';
+} from '../../../../extension-points/deployment-wizard';
 
 export type DeployExtension = ResolvedExtension<ModelServingDeploy<Deployment>>['properties'];
