+++ packages/model-serving/src/components/deploymentWizard/dynamicFormUtils.ts
@@ -1,34 +1,17 @@
 import React from 'react';
 import { useResolvedExtensions } from '@odh-dashboard/plugin-core';
 import type { RecursivePartial } from '@odh-dashboard/internal/typeHelpers';
-import type { DeploymentWizardField, WizardField, WizardFormData } from './types';
+import type { DeploymentWizardFieldOverride, WizardField, WizardFormData } from './types';
 import {
-  isDeploymentWizardFieldExtension,
-  isWizardField2Extension,
-} from '../../../extension-points';
+  isDeploymentWizardFieldOverrideExtension,
+  isWizardFieldExtension,
+} from '../../../extension-points/deployment-wizard';
 
-/**
- * @deprecated Use useWizardFieldOverrides instead
- */
-export const useWizardFieldFromExtension = <T extends DeploymentWizardField>(
-  predicate: (field: DeploymentWizardField) => field is T,
-  formData: RecursivePartial<WizardFormData['state']>,
-): T | undefined => {
-  const [extensions] = useResolvedExtensions(isDeploymentWizardFieldExtension);
-
-  const fields = React.useMemo(() => {
-    const all = extensions.map((ext) => ext.properties.field);
-    return all.filter(predicate).filter((field) => field.isActive(formData));
-  }, [extensions, predicate, formData]);
-
-  return React.useMemo(() => fields[0], [fields]);
-};
-
-export const useWizardFieldOverrides = <T extends DeploymentWizardField>(
-  predicate: (field: DeploymentWizardField) => field is T,
+export const useWizardFieldOverrides = <T extends DeploymentWizardFieldOverride>(
+  predicate: (field: DeploymentWizardFieldOverride) => field is T,
   formData: RecursivePartial<WizardFormData['state']>,
 ): T[] => {
-  const [extensions] = useResolvedExtensions(isDeploymentWizardFieldExtension);
+  const [extensions] = useResolvedExtensions(isDeploymentWizardFieldOverrideExtension);
 
   return React.useMemo(() => {
     const all = extensions.map((ext) => ext.properties.field);
@@ -39,7 +22,7 @@ export const useWizardFieldOverrides = <T extends DeploymentWizardField>(
 export const useActiveFields = (
   formData: RecursivePartial<WizardFormData['state']>,
 ): WizardField<unknown>[] => {
-  const [extensions] = useResolvedExtensions(isWizardField2Extension);
+  const [extensions] = useResolvedExtensions(isWizardFieldExtension);
 
   return React.useMemo(() => {
     return extensions
