+++ packages/model-serving/src/components/deploymentWizard/yaml/__tests__/useYamlResourcesResult.spec.ts
@@ -2,11 +2,8 @@ import { act } from '@testing-library/react';
 import { testHook } from '@odh-dashboard/jest-config/hooks';
 import { stringify } from 'yaml';
 import type { WizardFormData } from '../../types';
-import type {
-  Deployment,
-  ModelResourceType,
-  DeploymentAssemblyResources,
-} from '../../../../../extension-points';
+import type { Deployment, ModelResourceType } from '../../../../../extension-points';
+import { DeploymentAssemblyResources } from '../../../../../extension-points/deployment-wizard';
 import { useFormToResourcesTransformer } from '../useFormToResourcesTransformer';
 import { useFormYamlResources } from '../useYamlResourcesResult';
 import { useWizardFieldApply } from '../../useWizardFieldApply';
