+++ packages/model-serving/src/concepts/useCanMakeNewDeployment.ts
@@ -2,7 +2,7 @@ import { ProjectKind } from '@odh-dashboard/internal/k8sTypes';
 import { useKueueConfiguration } from '@odh-dashboard/internal/concepts/hardwareProfiles/kueueUtils';
 import { useExtensions } from '@odh-dashboard/plugin-core';
 import { useServingRuntimeTemplates } from './servingRuntimeTemplates/useServingRuntimeTemplates';
-import { isModelServingDeploy } from '../../extension-points';
+import { isModelServingDeploy } from '../../extension-points/deployment-wizard';
 
 export const useCanMakeNewDeployment = (
   project?: ProjectKind | null,
